requirements.txt
httpx==0.27.0
tenacity==8.3.0
playwright==1.44.0
anthropic==0.28.0
mysql-connector-python==8.4.0
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
loguru==0.7.2
